requirements.txt
pandas
requests
newsapi-python
openpyxl